src/requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy>=1.26.0,<2.0.0
plotly==5.18.0
scikit-learn>=1.3.0,<1.6.0
python-dotenv==1.0.0
requests==2.31.0
joblib>=1.3.0
edge-tts>=6.1.0
pygame>=2.5.0